chore(features): Remove obsolete systemsNotReporting flag
