+++ src/constants.js
@@ -234,7 +234,6 @@ export const COMPLIANCE_TABLE_DEFAULTS = {
 };
 
 export const features = {
-  systemsNotReporting: true,
   rbac: true,
   valueEditing: true,
   ruleGroups: true,
